requirements.txt
agno>=0.5.0
python-dotenv>=1.0.0
google-generativeai>=0.4.0